refactor(task): implementando uma verificação de ID do usuário no método deleteById de TaskService

+++ src/main/java/com/felipemarques/desafioNTIST/services/TaskService.java
@@ -49,6 +49,13 @@ public class TaskService {
     }
 
     public void deleteById(UUID taskId) {
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Task task = taskRepository.findByIdAndUserId(taskId, user.getId());
+
+        if(task == null) {
+            throw new TaskNotBelongToUser("A tarefa não pertence ao usuário logado!");
+        }
+
         taskRepository.deleteById(taskId);
     }
